+++ src/decorators/flatten.ts
@@ -1,9 +1,9 @@
-import { count } from "../internal/count";
+import {flatten} from "../internal/flatten";
 
-export function Count<T>(depth: number = Infinity) {
+export function Flatten<T>(depth: number = Infinity) {
     return (target: any, propertyKey: string) => {
-        let value: T[];
-        const getter = (): number => count(value, depth);
+        let value: unknown[];
+        const getter = (): T[] => flatten<T>(value, depth);
         const setter = (array: T[]) => value = array;
         Object.defineProperty(target, propertyKey, {
             get: getter,